% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{completeFun}
\alias{completeFun}
\title{Drop NAs by Columns}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{data}{data.frame object of observations}

\item{desiredCols}{list of columns from which to drop incomplete cases by}
}
\value{
Data frame with removed observations
}
\description{
Remove NAs based on specified columns in the data
}
\examples{
data <- data.frame(a=1:4,b=c("a","b","c","d"),c=c(NA,"keep",NA,"keep"))
completeFun(data, c("c"))

}
