% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{factorial}
\alias{factorial}
\title{Factorial}
\usage{
factorial(x)
}
\arguments{
\item{x}{Integer}
}
\value{
Factorial integer
}
\description{
Function that calculates factorial of integer
Note: this function is recursive!
}
\examples{
x <- 13
factorial(x)

}
