% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{bool, should vector be removed?}
}
\value{
float
}
\description{
Function to calculate geometric mean
}
\examples{
x <- c(1.4,1.6,NA,5.2,6.5,6.5)
gm_mean(x,na.rm=T)

}
