% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience_functions.R
\name{nonUnique}
\alias{nonUnique}
\title{Non-Unique}
\usage{
nonUnique(x)
}
\arguments{
\item{x}{vector/list}
}
\value{
vector/list
}
\description{
Function that returns a list of non-unique values from a supplied list
}
\examples{
x <- c(1,1,3,5,6,6)
nonUnique(x)

}
